Unify flow recovery notification types and decouple Telegram recovery from pause config.
- Canonical FlowNotificationIntentRecord in flows store with to_dict; ticket_flow_recovery aliases it for API stability.
- Add core flows recovery_notification_intents helpers for listing and marking delivery; reuse in Telegram and Discord adapters.
- Run Telegram recovery scans whenever the pause watcher runs, not only when pause notifications are enabled.
